requirements.txt
Django==3.1.1
django-crispy-forms==1.9.2
Pillow==7.2.0
django-taggit==1.3.0
